Add Rumor serialization and respective unit tests
Rumor uses the ISerializable interface to implement serialization. There
is one caveat to this implementation: Nodes may not assume that yields
will work across saves.

For example, assume there is a node that increments the variable foo and
then yields for an advance event. If the state of the game is saved during
that yield and the game is then resumed, foo will be incremented again
instead of resuming directly at the yield.