Rumor resets the state properly
The state needs to be reset at both the beginning and the end of the
execution. The reset at the beginning ensures that the state is clean
before execution starts. The reset at the end is necessary to get rid
of any existing state so that calls to the state after the script has
finished execution do not return invalid information.