Add support for specifying a speaker by variable
Support has been added by detecting whether or not the next token in the
say or add nodes is a double quote or not. If it is a double quote, then
it is assumed there is no defined speaker, and if it is not a double
quote, then it is assumed the speaker is defined by a single variable.

This is not exactly the desired behaviour as Rumor is able to accept more
flexible representations of speakers and speech. TODO comments have been
made in the relevant areas in the RumorCompiler where this issue can be
fixed.