Add Usage section to README
The example contained in the Usage section is based off of the
RumorScriptExample.cs file, but cannot be directly copy-pasted into
Unity due to the way paths work. This was the original reason for not
including a usage section.

However, it is much more convenient to see the potential usage of a
library right away, even if the example is not immediately usable as-is.