Fix precedence evaluation
Expressions like $ a = b + 1
didn't work because it split into +(=(a, b), 1) and evaluated the = before the +.
After this change, it splits by the operator with the largest index value first, so it generates the tree =(a, +(b, 1)) which means the + is evaluated on the b and 1 first.
The logic operators are also moved so that expressions like if a == 1 and b == 3: work as expected (as if it was (a==1) and (b==3) ).