Fix parser to allow functions and variables starting with keywords
This allows functions like add_something() or variables like choiceIsGood.
All keywords except else will have a space after them, and so we can match based on that. Unfortunately, else has a colon after it, so it needs a special case. Operators might not have a space after them, so we just match the operator without worrying about what's after it.