Fix scope serialization in .NET 4.x
Scope serialization doesn't work when the .NET 4.x runtime is selected.
This is because `GetValue` doesn't guarantee that the values are
deserialized until the deserialization constructor has exited. Instead,
we need to save the deserialized variables into instance variables, then
use them in a method marked with the `OnDeserialized` attribute to
repopulate the Dictionary.

The instance variables are then set to `null` to give the garbage
collector an opportunity to free up some memory.

Fixes #69.