Remove redundant OnClear messages
OnClear was invoked whenever `Rumor` was started or stopped. This
leads to some confusing behavior when relying heavily on callbacks,
as a listener may assume the OnClear callback is only invoked when a
clear command is encountered.

Additionally, OnClear was also invoked when a `jump` or `call` command
was encountered, which was incorrect. Additional tests have been added
to test these specific cases.