Remove requirement for `:`
The `:` character was originally used to denote that an indented block
of text must follow, but writers find it difficult to remember to do and
it is possible for no indented block of text to follow. For example,
a writer might desire a situation where the user is prompted to pick
a choice, but the choice has no impact.

In a future release, the trailing `:` character should be treated as an
error.