Add Transaction class
This allows the code to be more explicit about which parsers need to
succeed in sequence in order for the parser to succeed. This required
changing State to be an immutable class, and some parsers now get stuck
in an infinite loop.

This is likely because some parsers do not set their state properly when
exiting. I will attempt to address this problem by adding tests that
check the end state of the parsers.