Add user parser state to parser state
It doesn't make a lot of sense for the ParserState class to have a
hashset of all used identifiers, since it lives in the part of the
project which shouldn't know anything about the compiler.

Instead, the ParserState class can now contain an arbitrary state
defined by the user. I opted to use an interface to represent the type
of the user state instead of another type variable because using a type
variable would make the code a lot more difficult to read, especially
since we would need to also add a type constraint to almost every
parsing function just so we could call a method to clone the user state
for transactions.